+++ src/components/DropDownMenu/DropDownMenu.stories.tsx
@@ -19,13 +19,13 @@ const Desc = styled.div`
 
 const Description = () => (
   <Desc>
-    DEPRECATED используйте DropMenu. <br />
+    Deprecated используйте DropMenu. <br />
     Компонент Dropdown Menu имеет три размера и может быть с иконкой или без. Высота строки : xl - 48px, m - 40px, s -
     32px
   </Desc>
 );
 export default {
-  title: 'Deprecated/DropDownMenu (DEPRECATED используйте DropMenu)',
+  title: 'Deprecated/DropDownMenu (Deprecated используйте DropMenu)',
   decorators: [withDesign],
   component: DropDownMenu,
   parameters: {