+++ src/components/Tree/Tree.stories.tsx
@@ -334,7 +334,12 @@ const Demo2: ComponentStory<typeof Tree> = (props) => {
 
     return acc;
   }, {} as Record<any, any>);
-  const [activeItem, setActiveItem] = useState<string | undefined>('1');
+  const [activeItem, setActiveItem] = useState<string | null>('1');
+
+  const handleMouseLeave = () => {
+    // eslint-disable-next-line no-console
+    console.log('Mouse leave tree');
+  };
 
   return (
     <>
@@ -345,6 +350,7 @@ const Demo2: ComponentStory<typeof Tree> = (props) => {
         model={demo2_TreeModel}
         active={activeItem}
         onActivateItem={setActiveItem}
+        onMouseLeave={handleMouseLeave}
       />
     </>
   );