+++ src/components/input/FileUploader/FileUploader.stories.tsx
@@ -6,7 +6,7 @@ import { FileUploader } from '#src/components/input';
 import type { FileProps } from '#src/components/input/FileUploader/FileInfo';
 import { FileInfo } from '#src/components/input/FileUploader/FileInfo';
 import { T } from '#src/components/T';
-import type { Status } from './utils';
+import type { FileUploadStatus } from './utils';
 import { TextButton } from '#src/components/TextButton';
 import { ReactComponent as UploadSVG } from '@admiral-ds/icons/build/system/UploadOutline.svg';
 import type { Theme } from '#src/components/themes';
@@ -38,7 +38,7 @@ const Description = () => (
 );
 
 export default {
-  title: 'Deprecated/FileUploader (DEPRECATED используйте FileInput)',
+  title: 'Deprecated/FileUploader (Deprecated используйте FileInput)',
   decorators: [withDesign],
   component: FileUploader,
   parameters: {
@@ -149,7 +149,7 @@ const FileUploaderWithStatus: ComponentStory<typeof FileUploader> = (props) => {
 
   const [files, setFiles] = React.useState<FileProps[]>([]);
   const [isLoading, setIsLoading] = React.useState(false);
-  const statusOptions: Status[] = ['Loading', 'Queue', 'Uploaded', 'Error'];
+  const statusOptions: FileUploadStatus[] = ['Loading', 'Queue', 'Uploaded', 'Error'];
   const previewOptions = [true, false];
 
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {